adds command option to do no dependency checking, allow addition of new dependencies or do strict dependency checking
