+++ lib/src/diff/dependency_check_mode.dart
@@ -1,0 +1,5 @@
+enum DependencyCheckMode {
+  none,
+  allowAdding,
+  strict,
+}
