+++ lib/src/diff/diff.dart
@@ -1,1 +1,6 @@
+export 'api_change_type.dart';
+export 'api_change.dart';
+export 'dependency_check_mode.dart';
+export 'package_api_diff_result.dart';
+export 'package_api_differ_options.dart';
 export 'package_api_differ.dart';
