+++ lib/src/model/model.dart
@@ -1,9 +1,10 @@
 export 'declaration.dart';
-export 'package_api.dart';
-export 'interface_declaration.dart';
 export 'executable_declaration.dart';
 export 'field_declaration.dart';
-export 'type_alias_declaration.dart';
+export 'interface_declaration.dart';
+export 'package_api_semantics.dart';
+export 'package_api.dart';
+export 'package_dependency.dart';
 export 'platform_constraints.dart';
 export 'sdk_type.dart';
-export 'package_dependency.dart';
+export 'type_alias_declaration.dart';
